+++ bocadillo/recipes.py
@@ -1,22 +1,10 @@
-from typing import Sequence, Tuple, Any
+from typing import Sequence
+
 
 from .meta import DocsMeta
+from .misc import overrides
+from .routing import HTTPRoute, RoutingMixin, WebSocketRoute
 from .templates import TemplatesMixin
-from .routing import BaseRouter
-
-
-RecipeRoute = Tuple[Any, str, dict]
-
-
-class RecipeRouter(BaseRouter[RecipeRoute]):
-    def __init__(self):
-        super().__init__()
-        self.routes = []
-
-    def add_route(self, view: Any, pattern: str, **kwargs: Any) -> RecipeRoute:
-        route = (view, pattern, kwargs)
-        self.routes.append(route)
-        return route
 
 
 class RecipeBase:
@@ -38,7 +26,7 @@ class RecipeBase:
         raise NotImplementedError
 
 
-class Recipe(TemplatesMixin, RecipeBase, metaclass=DocsMeta):
+class Recipe(TemplatesMixin, RoutingMixin, RecipeBase, metaclass=DocsMeta):
     """A grouping of capabilities that can be merged back into an API.
 
     # Parameters
@@ -57,42 +45,30 @@ class Recipe(TemplatesMixin, RecipeBase, metaclass=DocsMeta):
             prefix = f"/{name}"
         super().__init__(prefix=prefix, **kwargs)
         self.name = name
-        self.http_router = RecipeRouter()
-        self.websocket_router = RecipeRouter()
 
-    def route(self, pattern: str, **kwargs):
-        """Register a route on the recipe.
+    def get_template_globals(self):
+        return {"url_for": self.url_for}
 
-        Accepts the same arguments as `API.route()`, except `namespace` which
-        will be given the value of the recipe's `name`.
-
-        # See Also
-        - [API.route()](./api.md#route)
-        """
+    @overrides(RoutingMixin.route)
+    def route(self, pattern: str, **kwargs) -> HTTPRoute:
         kwargs["namespace"] = self.name
-        return self.http_router.route(pattern=pattern, **kwargs)
+        return super().route(self.prefix + pattern, **kwargs)
 
-    def websocket_route(self, pattern: str, **kwargs):
-        """Register a WebSocket route on the recipe.
+    @overrides(RoutingMixin.websocket_route)
+    def websocket_route(self, pattern: str, **kwargs) -> WebSocketRoute:
+        return super().websocket_route(self.prefix + pattern, **kwargs)
 
-        Accepts the same arguments as `API.websocket_route()`.
-
-        # See Also
-        - [API.websocket_route()](./api.md#websocket-route)
+    def __call__(self, obj, root: str = ""):
+        """Apply the recipe to an object.
+        
+        Said object must be a subclass `RoutingMixin` and `TemplateMixin`.
         """
-        return self.websocket_router.route(pattern=pattern, **kwargs)
-
-    def __call__(self, api, root: str = ""):
-        """Apply the recipe to an API object."""
-        # Apply routes on the API
-        for view, pattern, kwargs in self.http_router.routes:
-            api.route(root + self.prefix + pattern, **kwargs)(view)
-        for view, pattern, kwargs in self.websocket_router.routes:
-            api.websocket_route(root + self.prefix + pattern, **kwargs)(view)
+        obj.http_router.mount(self.http_router, root=root)
+        obj.websocket_router.mount(self.websocket_router, root=root)
 
         # Look for templates where the API does, if not specified
         if self.templates_dir is None:
-            self.templates_dir = api.templates_dir
+            self.templates_dir = obj.templates_dir
 
     @classmethod
     def book(cls, *recipes: "Recipe", prefix: str) -> "RecipeBook":
@@ -116,7 +92,7 @@ class RecipeBook(RecipeBase):
         super().__init__(prefix)
         self.recipes = recipes
 
-    def __call__(self, api, root: str = ""):
+    def __call__(self, api: RoutingMixin, root: str = ""):
         """Apply the recipe book to an API object."""
         for recipe in self.recipes:
             recipe(api, root=root + self.prefix)